requirements.txt
discord.py
Flask
python-dotenv
Pillow
aiohttp
google-genai
